+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Expositum/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Expositum/releases)
 
-- Added: TOC version for patch 5.5.4 [mist of pandaria - classic]
-- Removed: TOC version for patch 12.0.1 [retail]
+- Added: Profile settings submenu with account-wide and character profile selection
+- Added: TOC version for patch 12.0.7 [retail]